refactor(#154): extract AbstractFreemarkerRenderer base class
Move the shared Freemarker boilerplate (Configuration injection, processTemplate
helper, and CDI-friendly setter) from VerifyRenderer and HkpIndexRenderer into a
common abstract base class AbstractFreemarkerRenderer.

Each renderer now only contains its own domain logic: model building and template
name selection. No behaviour change; all 23 tests pass.
